dom/media/tests/mochitest/test_zmedia_cleanup.html

<!DOCTYPE HTML>
<html>
  <head>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <script src="network.js"></script>
  </head>
<body>
<pre id="test">
<script type="application/javascript">

SimpleTest.waitForExplicitFinish();

if ("nsINetworkInterfaceListService" in SpecialPowers.Ci) {
  getNetworkUtils().tearDownNetwork()
    .then(() =>
          ok(true, 'Successfully teared down network interface'),
          () =>
          ok(true, 'Network interface was in down state already'))
    .then(() => SimpleTest.finish());
} else {
  ok(true, 'No need to cleanup network interface');
  SimpleTest.finish();
}

</script>
</pre>
</body>
</html>
